Extract route markers into shared widget module
